+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This is a small bugfix release.
 
 - Feat: add pl-PL translation
 - Upd: remove zh-Hant translation
+- Upd: IT translation
 - Fix: JS error when clearing the search box #982
 - Fix: update clear button for inputs on breadcrumb select
 - Fix: Finishing IME composition with Enter key also clears the search box #977