+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This is a small bugfix release.
 - Fix: JS error when clearing the search box #982
 - Fix: update clear button for inputs on breadcrumb select
 - Fix: Finishing IME composition with Enter key also clears the search box #977
+- Fix: Album view does not work with disabled Albumartist tag #988
 
 ***